spl-automation-main/requirement.txt
asgiref==3.5.2
Django==3.2.13
django-cors-headers==3.13.0
django-storages==1.12.3
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
typing-extensions==4.2.0
